pyproject.toml
[project]
name = "pyspark-livy-runner"
version = "0.1.0"
description = "Reusable Livy session launcher for PySpark"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "requests>=2.32",
  "pyyaml>=6.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]